fix: refac + fix fate core
