+++ lib/constants/showdownConverter.ts
@@ -1,0 +1,3 @@
+import showdown from "showdown";
+
+export const showdownConverter = new showdown.Converter();
