+++ lib/routes/SrdRoute/constants/SrdImport.tsx
@@ -1,0 +1,11 @@
+export const SrdImport = {
+  Condensed: () => {
+    return import("../../../srds/Fate-Condensed-SRD-CC-BY.md");
+  },
+  Core: () => {
+    return import("../../../srds/fate-core.md");
+  },
+  Accelerated: () => {
+    return import("../../../srds/fate-accelerated-SRD.md");
+  },
+};
